+++ prompt2model/run_locally.py
@@ -8,9 +8,9 @@ from prompt2model.demo_creator.gradio_creator import create_gradio
 from prompt2model.evaluator import MockEvaluator
 from prompt2model.model_executor import MockModelExecutor
 from prompt2model.model_retriever import MockModelRetriever
+from prompt2model.model_trainer import MockTrainer
 from prompt2model.param_selector import MockParamSelector
 from prompt2model.prompt_parser import DefaultSpec, PromptSpec, TaskType
-from prompt2model.trainer import MockTrainer
 
 parser = argparse.ArgumentParser()
 parser.add_argument(
