--- prompt2model/trainer/__init__.py
@@ -1,5 +1,0 @@
-"""Import ModelTrainer classes."""
-from prompt2model.trainer.base import ModelTrainer
-from prompt2model.trainer.mock import MockTrainer
-
-__all__ = ("MockTrainer", "ModelTrainer")
